Move views & urls to www/dashboard_favorites
