+++ lemarche/templates/dashboard/home_buyer.html
@@ -199,7 +199,7 @@
                         </div>
                     </div>
                     <div class="card-footer pt-0 bg-white text-right">
-                        <a href="{% url 'dashboard:profile_favorite_list' %}" id="track_dashboard_favorites" class="btn btn-link btn-ico">
+                        <a href="{% url 'dashboard_favorites:list' %}" id="track_dashboard_favorites" class="btn btn-link btn-ico">
                             <span>Voir mes favoris</span>
                             <i class="ri-arrow-right-s-line ri-xl"></i>
                         </a>
