+++ lemarche/templates/favorites/_favorite_item_remove_modal.html
@@ -7,7 +7,7 @@
                     <i class="ri-close-line"></i>
                 </button>
             </div>
-            <form method="POST" action="{% url 'dashboard:profile_favorite_item_delete' favorite_list.slug|default:'favorite-list-slug' siae.slug|default:'siae-slug-to-replace' %}">
+            <form method="POST" action="{% url 'dashboard_favorites:item_delete' favorite_list.slug|default:'favorite-list-slug' siae.slug|default:'siae-slug-to-replace' %}">
                 {% csrf_token %}
                 <div class="modal-body home-content-body">
                     <p class="font-weight-bold">
