+++ lemarche/templates/favorites/_favorite_list_delete_modal.html
@@ -7,7 +7,7 @@
                     <i class="ri-close-line"></i>
                 </button>
             </div>
-            <form method="POST" action="{% url 'dashboard:profile_favorite_list_delete' favorite_list.slug %}">
+            <form method="POST" action="{% url 'dashboard_favorites:list_delete' favorite_list.slug %}">
                 {% csrf_token %}
                 <div class="modal-body home-content-body">
                     <p class="font-weight-bold">Voulez vous supprimer cette liste d'achat ?</p>
