+++ lemarche/templates/favorites/_favorite_list_edit_modal.html
@@ -7,7 +7,7 @@
                     <i class="ri-close-line"></i>
                 </button>
             </div>
-            <form method="POST" action="{% url 'dashboard:profile_favorite_list_edit' favorite_list.slug %}">
+            <form method="POST" action="{% url 'dashboard_favorites:list_edit' favorite_list.slug %}">
                 {% csrf_token %}
                 <div class="modal-body home-content-body">
                     <p class="font-weight-bold">Modifier le nom de la liste</p>
