+++ lemarche/templates/favorites/dashboard_favorite_list_detail.html
@@ -12,7 +12,7 @@
                     <ol class="breadcrumb">
                         <li class="breadcrumb-item"><a href="{% url 'wagtail_serve' '' %}">Accueil</a></li>
                         <li class="breadcrumb-item"><a href="{% url 'dashboard:home' %}">Tableau de bord</a></li>
-                        <li class="breadcrumb-item"><a href="{% url 'dashboard:profile_favorite_list' %}">Liste d'achat favoris</a></li>
+                        <li class="breadcrumb-item"><a href="{% url 'dashboard_favorites:list' %}">Liste d'achat favoris</a></li>
                         <li class="breadcrumb-item active" aria-current="page">{{ favorite_list.name }}</li>
                     </ol>
                 </nav>
