+++ lemarche/www/dashboard/tests.py
@@ -4,7 +4,6 @@ from django.test import TestCase
 from django.urls import reverse
 from django.utils import timezone
 
-from lemarche.favorites.factories import FavoriteListFactory
 from lemarche.networks.factories import NetworkFactory
 from lemarche.siaes.factories import SiaeFactory
 from lemarche.siaes.models import SiaeUser
@@ -263,38 +262,6 @@ class DashboardSiaeUserViewTest(TestCase):
         self.assertEqual(response.url, reverse("dashboard:siae_users", args=[self.siae_with_users.slug]))
         self.assertEqual(self.siae_with_users.users.count(), 1)
         # TODO: user should not be able to delete itself
-
-
-class DashboardFavoriteListViewTest(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.user_favorite_list = UserFactory(kind=User.KIND_BUYER)
-        cls.other_user_favorite_list = UserFactory(kind=User.KIND_PARTNER)
-        cls.favorite_list_1 = FavoriteListFactory(name="Liste 1", user=cls.user_favorite_list)
-        cls.favorite_list_2 = FavoriteListFactory(name="Liste 2", user=cls.other_user_favorite_list)
-
-    def test_anonymous_user_cannot_view_favorite_list(self):
-        url = reverse("dashboard:profile_favorite_list")
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 302)
-        self.assertTrue(response.url.startswith("/accounts/login/"))
-        url = reverse("dashboard:profile_favorite_list_detail", args=[self.favorite_list_1.slug])
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 302)
-        self.assertTrue(response.url.startswith("/accounts/login/"))
-
-    def test_only_favorite_list_user_can_view_favorite_list_detail(self):
-        # non favorite list user
-        self.client.force_login(self.other_user_favorite_list)
-        url = reverse("dashboard:profile_favorite_list_detail", args=[self.favorite_list_1.slug])
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, "/profil/")
-        # favorite list user
-        self.client.force_login(self.user_favorite_list)
-        url = reverse("dashboard:profile_favorite_list_detail", args=[self.favorite_list_1.slug])
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200)
 
 
 class DashboardNetworkViewTest(TestCase):
