+++ lemarche/www/dashboard/views.py
@@ -8,11 +8,16 @@ from django.urls import reverse_lazy
 from django.utils import timezone
 from django.utils.safestring import mark_safe
 from django.views.generic import DeleteView, DetailView, ListView, UpdateView
-from django.views.generic.edit import CreateView, FormMixin
+from django.views.generic.edit import FormMixin
 
 from lemarche.cms.models import ArticlePage
 from lemarche.cms.snippets import ArticleCategory
-from lemarche.favorites.models import FavoriteItem, FavoriteList
+from lemarche.common.mixins import (
+    NetworkMemberRequiredMixin,
+    SiaeMemberRequiredMixin,
+    SiaeUserAndNotMemberRequiredMixin,
+    SiaeUserRequiredMixin,
+)
 from lemarche.networks.models import Network
 from lemarche.siaes.models import Siae, SiaeUser, SiaeUserRequest
 from lemarche.tenders.models import Tender, TenderSiae
@@ -20,7 +25,6 @@ from lemarche.users.models import User
 from lemarche.utils.s3 import S3Upload
 from lemarche.www.dashboard.forms import (
     ProfileEditForm,
-    ProfileFavoriteEditForm,
     SiaeClientReferenceFormSet,
     SiaeEditContactForm,
     SiaeEditInfoForm,
@@ -33,13 +37,6 @@ from lemarche.www.dashboard.forms import (
     SiaeSearchAdoptConfirmForm,
     SiaeSearchBySiretForm,
     SiaeUserRequestForm,
-)
-from lemarche.www.dashboard.mixins import (
-    FavoriteListOwnerRequiredMixin,
-    NetworkMemberRequiredMixin,
-    SiaeMemberRequiredMixin,
-    SiaeUserAndNotMemberRequiredMixin,
-    SiaeUserRequiredMixin,
 )
 from lemarche.www.dashboard.tasks import (
     send_siae_user_request_email_to_assignee,
@@ -118,108 +115,6 @@ class ProfileEditView(LoginRequiredMixin, SuccessMessageMixin, UpdateView):
 
     def get_object(self):
         return self.request.user
-
-
-class ProfileFavoriteListView(LoginRequiredMixin, ListView):
-    # form_class = ProfileFavoriteEditForm
-    template_name = "favorites/dashboard_favorite_list.html"
-    queryset = FavoriteList.objects.all()
-    context_object_name = "favorite_lists"
-
-    def get_queryset(self):
-        qs = super().get_queryset()
-        qs = qs.by_user(user=self.request.user)
-        return qs
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["form"] = ProfileFavoriteEditForm()
-        return context
-
-
-class ProfileFavoriteListCreateView(LoginRequiredMixin, SuccessMessageMixin, CreateView):
-    form_class = ProfileFavoriteEditForm
-    # success_message = "Votre liste d'achat a été crée avec succès."
-    success_url = reverse_lazy("dashboard:profile_favorite_list")
-
-    def form_valid(self, form):
-        """Add the User to the FavoriteList."""
-        favorite_list = form.save(commit=False)
-        favorite_list.user = self.request.user
-        favorite_list.save()
-
-        messages.add_message(
-            self.request,
-            messages.SUCCESS,
-            self.get_success_message(form.cleaned_data),
-        )
-        return HttpResponseRedirect(self.success_url)
-
-    def get_success_message(self, cleaned_data):
-        return mark_safe(f"Votre liste d'achat <strong>{cleaned_data['name']}</strong> a été crée avec succès.")
-
-
-class ProfileFavoriteListDetailView(FavoriteListOwnerRequiredMixin, DetailView):
-    template_name = "favorites/dashboard_favorite_list_detail.html"
-    queryset = FavoriteList.objects.prefetch_related("siaes").all()
-    context_object_name = "favorite_list"
-
-
-class ProfileFavoriteListEditView(FavoriteListOwnerRequiredMixin, SuccessMessageMixin, UpdateView):
-    form_class = ProfileFavoriteEditForm
-    template_name = "favorites/_favorite_list_edit_modal.html"
-    success_message = "Votre liste d'achat a été modifiée avec succès."
-    # success_url = reverse_lazy("dashboard:profile_favorite_list_detail")
-
-    def get_object(self):
-        return get_object_or_404(FavoriteList, slug=self.kwargs.get("slug"))
-
-    def get_success_url(self):
-        return reverse_lazy("dashboard:profile_favorite_list_detail", args=[self.kwargs.get("slug")])
-
-
-class ProfileFavoriteListDeleteView(FavoriteListOwnerRequiredMixin, SuccessMessageMixin, DeleteView):
-    template_name = "favorites/_favorite_list_delete_modal.html"
-    model = FavoriteList
-    # success_message = "Votre liste d'achat a été supprimée avec succès."
-    success_url = reverse_lazy("dashboard:profile_favorite_list")
-
-    def get_success_message(self, cleaned_data):
-        return mark_safe(f"Votre liste d'achat <strong>{self.object.name}</strong> a été supprimée avec succès.")
-
-
-class ProfileFavoriteItemDeleteView(LoginRequiredMixin, SuccessMessageMixin, DeleteView):
-    # FavoriteListOwnerRequiredMixin  # doesn't work because we don't have the FavoriteList slug
-    template_name = "siaes/_favorite_item_remove_modal.html"
-    model = FavoriteItem
-    # success_message = "La structure a été supprimée de votre liste d'achat avec succès."
-    success_url = reverse_lazy("dashboard:profile_favorite_list_detail")
-
-    def get_object(self):
-        """
-        - there should theoretically be only 1 Siae per user's lists (an Siae cannot belong to other list)
-        - in the future it could be possible to add an Siae to multiple user lists
-        """
-        # try:
-        #     favorite_list = FavoriteList.objects.get(slug=self.kwargs.get("slug"))
-        #     siae = Siae.objects.get(slug=self.kwargs.get("siae_slug"))
-        #     return get_object_or_404(FavoriteItem, favorite_list=favorite_list, siae=siae)
-        # except:  # noqa
-        #     raise Http404
-        siae = Siae.objects.get(slug=self.kwargs.get("siae_slug"))
-        return get_object_or_404(FavoriteItem, favorite_list__user=self.request.user, siae=siae)
-
-    def get_success_url(self):
-        """Redirect to the previous page."""
-        request_referer = self.request.META.get("HTTP_REFERER", "")
-        if request_referer:
-            return request_referer
-        return super().get_success_url()
-
-    def get_success_message(self, cleaned_data):
-        return mark_safe(
-            f"<strong>{self.object.siae.name_display}</strong> a été supprimée de votre liste d'achat avec succès."
-        )
 
 
 class ProfileNetworkDetailView(NetworkMemberRequiredMixin, DetailView):
